+++ mnistdata.h
@@ -15,13 +15,13 @@ const unsigned int INPUT_SIZE = 784;
 const unsigned int OUTPUT_SIZE = 10;
 
 struct Training_Datum {
-  valarray<double> x;
+  valarray<float> x;
   int y;
-  valarray<double> t;
+  valarray<float> t;
 };
 typedef vector<Training_Datum> Training_Data ;
 
-valarray<double> to_model_output(int in) ;
+valarray<float> to_model_output(int in) ;
 Training_Data load_data_from_file(string filename, int cutoff) ;
 }
 #endif /* MNISTDATA_H */
